requirements.txt
numpy>=1.26,<2.0
PyYAML>=6.0
torch==2.7.0
gymnasium>=0.29.1
highway-env==1.10.2
stable-baselines3>=2.3.2
tqdm>=4.66
matplotlib>=3.8
imageio>=2.37
imageio-ffmpeg>=0.6
tensorboard>=2.16
